+++ packages/sdks/vue-sdk/src/UserProfile.vue
@@ -10,6 +10,7 @@
       :debug.attr="debug"
       :widget-id="widgetId"
       @logout="onLogout"
+      @ready="onReady"
     />
   </div>
 </template>
@@ -20,12 +21,17 @@ import { DESCOPE_INJECTION_KEY } from './constants';
 import '@descope/user-profile-widget';
 const descope = inject(DESCOPE_INJECTION_KEY);
 import { useOptions } from './hooks';
-const emit = defineEmits(['logout']);
+const emit = defineEmits<{
+  (e: 'logout', payload: Event): void;
+  (e: 'ready', payload: CustomEvent<Record<string, never>>): void;
+}>();
 
 const onLogout = (e: Event) => {
   descope?.resetAuth(); // reset session values
   emit('logout', e);
 };
+
+const onReady = (e: CustomEvent<Record<string, never>>) => emit('ready', e);
 
 defineProps({
   widgetId: {