+++ packages/widgets/audit-management-widget/src/lib/widget/mixins/initMixin/initComponentsMixins/initWidgetRootMixin.ts
@@ -41,7 +41,8 @@ export const initWidgetRootMixin = createSingletonMixin(
 
         await Promise.all([this.#initWidgetRoot()]);
 
-        this.onWidgetRootReady();
+        await this.onWidgetRootReady();
+        this.dispatchEvent(new CustomEvent('ready'));
       }
     },
 );