+++ packages/widgets/role-management-widget/src/lib/widget/mixins/initMixin/initComponentsMixins/initWidgetRootMixin.ts
@@ -43,7 +43,8 @@ export const initWidgetRootMixin = createSingletonMixin(
           this.actions.getTenantPermissions(),
         ]);
 
-        this.onWidgetRootReady();
+        await this.onWidgetRootReady();
+        this.dispatchEvent(new CustomEvent('ready'));
       }
     },
 );