+++ packages/libs/sdk-component-drivers/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.6.2](https://github.com/descope/descope-js/compare/sdk-component-drivers-0.6.1...sdk-component-drivers-0.6.2) (2025-09-28)
+
+### Dependency Updates
+
+* `sdk-helpers` updated to version `0.4.1`
 ## [0.6.1](https://github.com/descope/descope-js/compare/sdk-component-drivers-0.6.0...sdk-component-drivers-0.6.1) (2025-09-07)