+++ packages/libs/sdk-helpers/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.4.1](https://github.com/descope/descope-js/compare/sdk-helpers-0.4.0...sdk-helpers-0.4.1) (2025-09-28)
+
 ## [0.4.0](https://github.com/descope/descope-js/compare/sdk-helpers-0.3.0...sdk-helpers-0.4.0) (2025-08-14)