+++ packages/sdks/react-sdk/CHANGELOG.md
@@ -2,6 +2,44 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [2.21.1](https://github.com/descope/descope-js/compare/react-sdk-2.21.0...react-sdk-2.21.1) (2025-10-13)
+
+### Dependency Updates
+
+* `user-management-widget` updated to version `0.9.15`
+## [2.21.0](https://github.com/descope/descope-js/compare/react-sdk-2.20.5...react-sdk-2.21.0) (2025-09-29)
+
+### Dependency Updates
+
+* `access-key-management-widget` updated to version `0.5.14`
+* `audit-management-widget` updated to version `0.5.14`
+* `role-management-widget` updated to version `0.5.6`
+* `user-management-widget` updated to version `0.9.14`
+* `user-profile-widget` updated to version `0.6.22`
+* `applications-portal-widget` updated to version `0.4.14`
+* `web-component` updated to version `3.47.9`
+* `web-js-sdk` updated to version `1.37.0`
+* `tenant-profile-widget` updated to version `0.2.19`
+
+### Features
+
+* add custom cookie domain ([#1223](https://github.com/descope/descope-js/issues/1223)) RELEASE ([a3fd659](https://github.com/descope/descope-js/commit/a3fd6595c71ada954cce785962d814dde5cc4450))
+
+## [2.20.5](https://github.com/descope/descope-js/compare/react-sdk-2.20.4...react-sdk-2.20.5) (2025-09-28)
+
+### Dependency Updates
+
+* `sdk-helpers` updated to version `0.4.1`
+* `access-key-management-widget` updated to version `0.5.13`
+* `audit-management-widget` updated to version `0.5.13`
+* `role-management-widget` updated to version `0.5.5`
+* `user-management-widget` updated to version `0.9.13`
+* `user-profile-widget` updated to version `0.6.21`
+* `applications-portal-widget` updated to version `0.4.13`
+* `web-component` updated to version `3.47.8`
+* `web-js-sdk` updated to version `1.36.2`
+* `core-js-sdk` updated to version `2.50.1`
+* `tenant-profile-widget` updated to version `0.2.18`
 ## [2.20.4](https://github.com/descope/descope-js/compare/react-sdk-2.20.3...react-sdk-2.20.4) (2025-09-14)
 
 ### Dependency Updates