+++ packages/widgets/applications-portal-widget/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.4.14](https://github.com/descope/descope-js/compare/applications-portal-widget-0.4.13...applications-portal-widget-0.4.14) (2025-09-29)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.37.0`
+## [0.4.13](https://github.com/descope/descope-js/compare/applications-portal-widget-0.4.12...applications-portal-widget-0.4.13) (2025-09-28)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.36.2`
+* `sdk-helpers` updated to version `0.4.1`
+* `sdk-mixins` updated to version `0.13.13`
+* `sdk-component-drivers` updated to version `0.6.2`
 ## [0.4.12](https://github.com/descope/descope-js/compare/applications-portal-widget-0.4.11...applications-portal-widget-0.4.12) (2025-09-14)
 
 ### Dependency Updates